@@ -1,13 +1,13 @@
 package com.ibm.optim.oaas.sample.trucking;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.util.List;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import com.ibm.optim.oaas.sample.trucking.model.Hub;
 import com.ibm.optim.oaas.sample.trucking.model.Shipment;
@@ -16,83 +16,80 @@ import com.ibm.optim.oaas.sample.trucking.model.Spoke;
 import com.ibm.optim.oaas.sample.trucking.model.TruckType;
 
 public class TruckingIT {
+  private final String SERVER = "http://localhost:9080/";
 
-	final String SERVER = "http://localhost:9080/";
+  @Test
+  public void testShipments() throws Exception {
+    final TruckingClient client = new TruckingClient(SERVER);
 
-	@Test
-	public void testShipments() throws Exception {
-		TruckingClient client = new TruckingClient(SERVER);
+    client.initialize();
 
-		client.initialize();
+    final List<Shipment> shipments = client.getShipments();
+    assertNotNull(shipments);
+    assertEquals(shipments.size(), 30);
 
-		List<Shipment> shipments = client.getShipments();
-		assertNotNull(shipments);
-		assertEquals(shipments.size(), 30);
+    final Shipment s = new Shipment();
+    s.setOrigin("A");
+    s.setDestination("F");
+    s.setTotalVolume(500);
 
-		Shipment s = new Shipment();
-		s.setOrigin("A");
-		s.setDestination("F");
-		s.setTotalVolume(500);
+    final Shipment rs = client.addShipment(s);
+    assertNotNull(rs.getId());
+    assertEquals(s.getOrigin(), rs.getOrigin());
+    assertEquals(s.getDestination(), rs.getDestination());
+    assertEquals(s.getTotalVolume(), rs.getTotalVolume());
 
-		Shipment rs = client.addShipment(s);
-		assertNotNull(rs.getId());
-		assertEquals(s.getOrigin(), rs.getOrigin());
-		assertEquals(s.getDestination(), rs.getDestination());
-		assertEquals(s.getTotalVolume(), rs.getTotalVolume());
+    final Shipment gs = client.getShipment(rs.getId());
+    assertNotNull(gs);
+    assertEquals(rs.getOrigin(), gs.getOrigin());
+    assertEquals(rs.getDestination(), gs.getDestination());
+    assertEquals(rs.getTotalVolume(), gs.getTotalVolume());
 
-		Shipment gs = client.getShipment(rs.getId());
-		assertNotNull(gs);
-		assertEquals(rs.getOrigin(), gs.getOrigin());
-		assertEquals(rs.getDestination(), gs.getDestination());
-		assertEquals(rs.getTotalVolume(), gs.getTotalVolume());
+    client.deleteShipment(rs.getId());
 
-		client.deleteShipment(rs.getId());
+    final Shipment ds = client.getShipment(rs.getId());
+    assertNull(ds);
 
-		Shipment ds = client.getShipment(rs.getId());
-		assertNull(ds);
+    client.deleteShipments();
 
-		client.deleteShipments();
+    final List<Shipment> dshipments = client.getShipments();
+    assertNotNull(dshipments);
+    assertEquals(dshipments.size(), 0);
+  }
 
-		List<Shipment> dshipments = client.getShipments();
-		assertNotNull(dshipments);
-		assertEquals(dshipments.size(), 0);
+  @Test
+  public void testInitialize() throws Exception {
+    final TruckingClient client = new TruckingClient(SERVER);
 
-	}
+    client.initialize();
 
-	@Test
-	public void testInitialize() throws Exception {
-		TruckingClient client = new TruckingClient(SERVER);
+    final List<TruckType> truckTypes = client.getTruckTypes();
+    assertNotNull(truckTypes);
+    assertFalse(truckTypes.isEmpty());
 
-		client.initialize();
+    final List<Spoke> spokes = client.getSpokes();
+    assertNotNull(spokes);
+    assertFalse(spokes.isEmpty());
 
-		List<TruckType> truckTypes = client.getTruckTypes();
-		assertNotNull(truckTypes);
-		assertFalse(truckTypes.isEmpty());
+    final List<Hub> hubs = client.getHubs();
+    assertNotNull(hubs);
+    assertFalse(hubs.isEmpty());
 
-		List<Spoke> spokes = client.getSpokes();
-		assertNotNull(spokes);
-		assertFalse(spokes.isEmpty());
+    final List<Shipment> shipments = client.getShipments();
+    assertNotNull(shipments);
+    assertFalse(shipments.isEmpty());
+  }
 
-		List<Hub> hubs = client.getHubs();
-		assertNotNull(hubs);
-		assertFalse(hubs.isEmpty());
+  @Test
+  public void testSolve() throws Exception {
+    final TruckingClient client = new TruckingClient(SERVER);
 
-		List<Shipment> shipments = client.getShipments();
-		assertNotNull(shipments);
-		assertFalse(shipments.isEmpty());
-	}
+    client.initialize();
 
-	@Test
-	public void testSolve() throws Exception {
-		TruckingClient client = new TruckingClient(SERVER);
+    client.solve();
 
-		client.initialize();
-
-		client.solve();
-
-		Solution solution = client.getSolution();
-		assertNotNull(solution);
-		assertNotNull(solution.getJobid());
-
-	}
+    final Solution solution = client.getSolution();
+    assertNotNull(solution);
+    assertNotNull(solution.getJobid());
+  }
 }
